+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Remove `<nav>` from pagination (#686, @xi).
 - Add an `id` to the help text of fields for Django 5.0+, to match the `aria-describedby` attribute.
 - Drop support for Python 3.8 in the test matrix
+- Add field paramaters for custom valid and invalid classes
 
 ## 24.3 (2024-09-17)